Modify TensorFlow Saver examples to manually set name property
TensorFlow uses a string identifier for Tensors and Operations called name.
The name property is set for nodes automatically based on the order they
appear, so if you are restoring a previous session you may encounter
errors if the TensorFlow variables are defined in a different order. By
setting the names manually this potential issue is avoided.